TestWasapiDownloader split into tests requiring PowerMock and not (for better coverage info)
